requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
flask==2.3.3
gunicorn==21.2.0
firebase-admin==6.2.0
google-cloud-storage==2.10.0
google-cloud-firestore==2.11.1

# Gemini API
google-generativeai==0.3.1

# Media processing
ffmpeg-python==0.2.0
opencv-python==4.8.0.76
numpy==1.24.3
pillow==10.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
